feat: Create structure for Form component on page CreateTraining
